feat: refactor admin tests
